requirements.txt
langchain==0.3.27
langchain_community==0.3.29
langchain_openai==0.3.33
pydantic==2.11.9
python-dotenv==1.1.1
